+++ backend/src/app.module.ts
@@ -4,6 +4,7 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { UsersModule } from './users/users.module';
 import { User } from './users/user.model';
+import { LogsModule } from './_logs/logs.module';
 
 @Module({
   imports: [
@@ -20,6 +21,7 @@ import { User } from './users/user.model';
       synchronize: true,
     }),
     UsersModule,
+    LogsModule,
   ],
   controllers: [AppController],
   providers: [AppService],