+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Create module for log
 - Create frontend for the project
 - Create endpoint for adding, removing and getting users. And attach class diagram
 - Implement database connection